Dockerizing_a_simple_Python_Flask_web_application/requirement.txt
Flask==3.1.0
openai==0.27.10
genai==2.1.0
google-cloud==0.34.0
google-genai==1.3.0